+++ components/Screens/HeavyCoat/HeavyCoatForm.tsx
@@ -4,58 +4,24 @@ import { ReactElement, useContext, useState } from 'react';
 import { Button, Select } from 'sk-storybook';
 
 import * as S from './HeavyCoatForm.styles';
-import { useQuestionnaireNextScreenURL } from '../../../hooks/use-questionnaire-next-screen-url';
-import axios from 'axios';
-import { Dog } from '../../../entities/dog.entities';
-import { useRouter } from 'next/router';
-import UserContext from '../../../context/user.context';
+import { Option } from '../../../entities/questionnaire.entities';
 
-export type Option = {
-  label: string;
-  value: string | number | boolean;
-};
+interface HeavyCoatFormProps {
+  handleSubmit: (event: React.SyntheticEvent) => void;
+  setValue: React.Dispatch<React.SetStateAction<Option | undefined>>;
+  value: Option | undefined;
+}
 
 const options: Option[] = [
   { label: 'Yes', value: true },
   { label: 'No', value: false },
 ];
 
-export const HeavyCoatForm = (): ReactElement => {
-  const [value, setValue] = useState<Option | undefined>();
-  const userContext = useContext(UserContext);
-  const router = useRouter();
-  const dogId = router.query.dogId;
-
-  const handleSubmit = async (event: React.SyntheticEvent) => {
-    //TODO: Implement logic for onSubmit
-    event.preventDefault();
-    console.log(value?.value);
-
-    try {
-      if (userContext.user && value?.value !== undefined) {
-        await axios
-          .post('http://localhost:3001/api/dog/heavy-coat', {
-            dogId,
-            heavyCoat: value.value,
-            userId: userContext.user.id,
-          })
-          .then((res) => {
-            const dog: Dog = res.data;
-            const nextScreenUrl = useQuestionnaireNextScreenURL(dog);
-            router.push(nextScreenUrl);
-          })
-          .catch((error) => {
-            console.error('An error occurred:', error); //TODO: Handle error - Toast message
-          });
-      } else {
-        console.log('no user');
-      }
-    } catch (error) {
-      console.error('An error occurred:', error);
-      // TODO: Handle error - Toast message
-    }
-  };
-
+export const HeavyCoatForm = ({
+  handleSubmit,
+  setValue,
+  value,
+}: HeavyCoatFormProps): ReactElement => {
   return (
     <S.Container>
       <form onSubmit={(event) => handleSubmit(event)}>
